useEffect Hook:
Dependency Array: An empty array ([]) indicates that this effect runs only once after the initial render.
Effect Logic:
Condition: Checks if the user is null or if the user.role is not 'recruiter'.
If true, it means the user is either not logged in or does not have the correct role. In this case, it redirects the user to the root path ("/").
Purpose: Ensures that only users with the role of 'recruiter' can access the content wrapped by this ProtectedRoute component.